Simplified the templates directory, added writing prompts
It was confusing to have so many different files in the template
directory, both to authors and readers of the proposals. This
commit simplifies the directory of templates down to one 'index.md'
and fills it out with sections that contain writing prompts taken
from JEP-2 combined with Henrik's previous example text.