Updating Capabilities Armature
The main document was previously a long description of something
more like philosophy, and now it's a more clear description
of what I think we should do and why, with examples.

I've also removed my old description, as it was too free-form
and had some contradictory information in it.

Note that the Armature templates do not actually talk about content
itself.  They talk about some basics, but you can't take the templates
and usefully build an armature from them.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>