.epp should always be the preferred extension by convention
This is the recommended behavior with ERB. There are a handful of reasons for this, most of which being good editor support. Presumably support for the EPP syntax will need to be added to most common editors, and the `.epp` extension (or in this example `.html.epp`) is the easiest way to identify the syntax to use (rather than relying on magic comments or manual intervention).