Update ARM-8 for dynamic environment support
Currently, static environments may be created by adding custom
`[whatever]` sections to `puppet.conf`, and in the absense of a matching
environment, the defaults will be used. The defaults can be "dynamic" by
the incorporation of `$environment` in the `manifest` and `modulepath`
parameters. This pattern is very widely used for dynamic environment
creation.

This addition updates the armature to indicate how `Environment`
resource defaults could be used to enable dynamic environment support,
and adds a parameter `dynamic_environments` that may be used to disable
this behaviour.