Simplify ACL Armature
This removes ACEs as a type as they are nothing more than a data type, which
doesn't exist in puppet types.

t:This also munges security descriptor and acl into one so that ensure is not
confusing. This allows a future expansion to include linux acls as the type
structure could apply to multiple platforms.